# src/spec/workload-openapi-spec.yaml
openapi: 3.0.2
servers:
  - url: /api/v1
info:
  description: "API for the integration of Openshift Telemetry Analytics workload analyis results with any consumer."
  version: "1.0.0"
  title: "Openshift Telemetry Workload Analysis API"
...
paths:
...
  /readiness:
    get:
      tags:
        - "Health check"
      summary: "Check whether the service is ready to serve requests."
      description: "A readiness probe to check whether the service is ready."
      operationId: "routes.health_check.readiness"
      responses:
        "200":
          description: "service is ready"
        "500":
          description: "Service is down/internal exception."
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      x-bearerInfoFunc: routes.auth.decode_token
...